pipe Tests: long pipefd[2] → int pipefd[2] (Linux ABI)
pipe()/pipe2() schreibt int[2] (8 Bytes). Tests benutzten
long[2] (16 Bytes) → auf x86_64 little-endian wurden beide
int-Werte in pipefd[0] gepackt, pipefd[1] war uninitialisiert.

Resultat: write/read an falsche FDs (oft stdout/serial).
Betroffen: epoll_wait, epoll_ctl, accept, bind, close,
dup, fadvise, fcntl, fdatasync — 20+ Tests repariert.

+++ test/ltp/test_bind.c
@@ -103,7 +103,7 @@ static void test_bind04(void) {
     sc1(SYS_UNLINK, (long)"/tmp/bind04");
 
     /* Pipe */
-    long pipefd[2];
+    int pipefd[2];
     r = sc1(SYS_PIPE, (long)pipefd);
     check_val("pipe", r, 0);
     if (r == 0) {
